+++ packages/components-js/tests/smoke/unit/specs/cdn.smoke.ts
@@ -10,8 +10,6 @@ import {
   FONTS_MANIFEST,
   ICONS_CDN_BASE_PATH,
   ICONS_MANIFEST,
-  MARQUES_CDN_BASE_PATH,
-  MARQUES_MANIFEST,
   META_ICONS_CDN_BASE_PATH,
   META_ICONS_MANIFEST,
   MODEL_SIGNATURES_CDN_BASE_PATH,
@@ -83,11 +81,6 @@ describe('cdn', () => {
   describe('flags', () => {
     const flags = objectToFlatArray(FLAGS_MANIFEST);
     bulkRequestItems(flags, `${CDN_BASE_URL_COM}${FLAGS_CDN_BASE_PATH}`);
-  });
-
-  describe('marque', () => {
-    const marques = objectToFlatArray(MARQUES_MANIFEST);
-    bulkRequestItems(marques, `${CDN_BASE_URL_COM}${MARQUES_CDN_BASE_PATH}`);
   });
 
   describe('meta-icons', () => {