+++ packages/components/src/components/flag/flag-styles.ts
@@ -1,35 +1,14 @@
 import { addImportantToEachRule, hostHiddenStyles } from '../../styles';
-import {
-  fontPorscheNext,
-  leadingNormal,
-  typescale2Xl,
-  typescale2Xs,
-  typescaleLg,
-  typescaleMd,
-  typescaleSm,
-  typescaleXl,
-  typescaleXs,
-} from '../../styles/css-variables';
+import { fontPorscheNext, leadingNormal } from '../../styles/css-variables';
+import { sizeMap } from '../../styles/maps';
 import { buildResponsiveStyles, getCss } from '../../utils';
 import type { BreakpointCustomizable } from '../../utils/breakpoint-customizable';
-import type { IconSize } from '../icon/icon-utils';
 import type { FlagSize } from './flag-utils';
 
 /**
  * @css-variable {"name": "--p-flag-size", "description": "Defines the width and height of the flag. Overrides the `size` property when set.", "defaultValue": ""}
  */
 const cssVarSize = '--p-flag-size';
-
-const sizeMap: Record<FlagSize, string> = {
-  'xx-small': typescale2Xs,
-  'x-small': typescaleXs,
-  small: typescaleSm,
-  medium: typescaleMd,
-  large: typescaleLg,
-  'x-large': typescaleXl,
-  'xx-large': typescale2Xl,
-  inherit: 'inherit',
-};
 
 export const getComponentCss = (size: BreakpointCustomizable<FlagSize>): string => {
   const dimension = `var(${cssVarSize},${leadingNormal})`;
@@ -55,7 +34,7 @@ export const getComponentCss = (size: BreakpointCustomizable<FlagSize>): string 
         width: dimension,
         height: dimension,
         fontFamily: fontPorscheNext, // needed for correct width/height definition based on ex-unit
-        ...buildResponsiveStyles(size, (s: IconSize) => ({
+        ...buildResponsiveStyles(size, (s: FlagSize) => ({
           fontSize: sizeMap[s], // needed for correct width/height definition based on ex-unit
         })),
       },