+++ packages/components/src/components/heading/heading.tsx
@@ -1,21 +1,26 @@
 import { Component, Element, h, type JSX, Prop } from '@stencil/core';
-import type { BreakpointCustomizable, HeadingSize, HeadingTag, PropTypes } from '../../types';
+import type { BreakpointCustomizable, PropTypes } from '../../types';
+import { AllowedTypes, attachComponentCss, hasPropValueChanged, validateProps } from '../../utils';
+import { getComponentCss } from './heading-styles';
 import {
-  AllowedTypes,
-  attachComponentCss,
+  getHeadingTagType,
+  HEADING_ALIGNS,
+  HEADING_COLORS,
   HEADING_SIZES,
   HEADING_TAGS,
-  hasPropValueChanged,
-  TYPOGRAPHY_ALIGNS,
-  validateProps,
-} from '../../utils';
-import { getComponentCss } from './heading-styles';
-import { getHeadingTagType, HEADING_COLORS, type HeadingAlign, type HeadingColor } from './heading-utils';
+  HEADING_WEIGHTS,
+  type HeadingAlign,
+  type HeadingColor,
+  type HeadingSize,
+  type HeadingTag,
+  type HeadingWeight,
+} from './heading-utils';
 
 const propTypes: PropTypes<typeof Heading> = {
   tag: AllowedTypes.oneOf<HeadingTag>([undefined, ...HEADING_TAGS]),
   size: AllowedTypes.breakpoint<HeadingSize>(HEADING_SIZES),
-  align: AllowedTypes.oneOf<HeadingAlign>(TYPOGRAPHY_ALIGNS),
+  weight: AllowedTypes.oneOf<HeadingWeight>(HEADING_WEIGHTS),
+  align: AllowedTypes.oneOf<HeadingAlign>(HEADING_ALIGNS),
   color: AllowedTypes.oneOf<HeadingColor>(HEADING_COLORS),
   ellipsis: AllowedTypes.boolean,
 };
@@ -30,19 +35,22 @@ const propTypes: PropTypes<typeof Heading> = {
 export class Heading {
   @Element() public host!: HTMLElement;
 
-  /** Sets a heading tag, so it fits correctly within the outline of the page. */
+  /** Sets the HTML heading tag (h1 - h6) to ensure the correct document outline and semantic hierarchy. If not set, the tag is automatically inferred from the `size` property. */
   @Prop() public tag?: HeadingTag;
 
-  /** Size of the component. Also defines the size for specific breakpoints, like {base: "small", l: "medium"}. You always need to provide a base value when doing this. */
-  @Prop() public size?: BreakpointCustomizable<HeadingSize> = 'xx-large';
+  /** Size of the heading. Also defines the size for specific breakpoints, like {base: "md", l: "2xl"}. */
+  @Prop() public size?: BreakpointCustomizable<HeadingSize> = '2xl';
 
-  /** Text alignment of the component. */
+  /** The font weight of the heading. For `size` values of 'sm' or smaller, it's recommended to use 'semibold' for better readability. */
+  @Prop() public weight?: HeadingWeight = 'normal';
+
+  /** Text alignment of the heading. */
   @Prop() public align?: HeadingAlign = 'start';
 
-  /** Basic text color variations. */
+  /** Text color of the heading. Use 'primary' for default, 'contrast-high' / 'contrast-medium' for alternative emphasis, or 'inherit' to adopt the parent's color. */
   @Prop() public color?: HeadingColor = 'primary';
 
-  /** Adds an ellipsis to a single line of text if it overflows. */
+  /** Adds an ellipsis to a single line of text if it overflows the container width. */
   @Prop() public ellipsis?: boolean = false;
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
@@ -51,7 +59,7 @@ export class Heading {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    attachComponentCss(this.host, getComponentCss, this.size, this.align, this.color, this.ellipsis);
+    attachComponentCss(this.host, getComponentCss, this.size, this.weight, this.align, this.color, this.ellipsis);
 
     const TagType = getHeadingTagType(this.host, this.size, this.tag);