+++ packages/components/src/styles/maps/index.ts
@@ -1,0 +1,3 @@
+export { colorMap } from './color-map';
+export { sizeMap } from './size-map';
+export { weightMap } from './weight-map';