+++ packages/components/src/utils/index.ts
@@ -35,5 +35,4 @@ export * from './supportsNativePopover';
 export * from './sync';
 export { getPrefixedTagNames, getTagName, getTagNameWithoutPrefix } from './tag-name';
 export * from './tile';
-export * from './typography';
 export * from './validation';