+++ packages/storefront/src/app/stylesheets/color/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation';
+
+export default function Page() {
+  redirect('/stylesheets/color/examples');
+}