+++ packages/styles/projects/emotion/src/typography/proseHeading5XlStyle.ts
@@ -1,0 +1,5 @@
+import { fontPorscheNext, fontWeightNormal, leadingNormal, typescale5Xl } from '@porsche-design-system/tokens';
+
+export const proseHeading5XlStyle = {
+  font: `normal normal ${fontWeightNormal} ${typescale5Xl} / ${leadingNormal} ${fontPorscheNext}`,
+};