+++ packages/tokens/src/font/size/typescale4Xl.ts
@@ -1,0 +1,1 @@
+export const typescale4Xl = 'clamp(2.03rem, 3.58vw + 1.31rem, 5.61rem)';